SPOON-971 Add checking for zip extension and places file in local storage

+++ server/openstorefront/openstorefront-web/src/main/java/edu/usu/sdl/openstorefront/web/action/UploadAction.java
@@ -41,6 +41,7 @@ import edu.usu.sdl.openstorefront.security.SecurityUtil;
 import edu.usu.sdl.openstorefront.service.io.parser.MainAttributeParser;
 import edu.usu.sdl.openstorefront.service.io.parser.OldBaseAttributeParser;
 import edu.usu.sdl.openstorefront.service.manager.DBManager;
+import edu.usu.sdl.openstorefront.service.manager.MailManager;
 import edu.usu.sdl.openstorefront.validation.ValidationModel;
 import edu.usu.sdl.openstorefront.validation.ValidationResult;
 import edu.usu.sdl.openstorefront.validation.ValidationUtil;
@@ -49,6 +50,7 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.RandomAccessFile;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -73,7 +75,10 @@ import net.sourceforge.stripes.action.FileBean;
 import net.sourceforge.stripes.action.HandlesEvent;
 import net.sourceforge.stripes.action.Resolution;
 import net.sourceforge.stripes.validation.Validate;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang.StringUtils;
+import org.codemonkey.simplejavamail.email.Email;
+import javax.mail.Message;
 
 /**
  *
@@ -624,42 +629,50 @@ public class UploadAction
 		Map<String, String> errors = new HashMap<>();
 		LOG.log(Level.INFO, SecurityUtil.adminAuditLogMessage(getContext().getRequest()));
 
-		File tempFile = null;
-		try {
-			// save to file system under username
+		String extension = FilenameUtils.getExtension(uploadFile.getFileName()).toString();
+		String mimeType = uploadFile.getContentType();
+
+		Boolean isZip = extension.equals("zip") && (mimeType.equals("application/x-zip-compressed") ||
+		mimeType.equals("application/zip") ||
+		mimeType.equals("application/octet-stream") ||
+		mimeType.equals("multipart/x-zip"));
+
+		if(isZip){
+			String username = SecurityUtil.getCurrentUserName();
+
+			File tempFile = null;
 			String timeStamp = new SimpleDateFormat("dd-MM-YYYY").format(new Date());
-			tempFile = new File(FileSystemManager.BULK_UPLOAD_DIR + "/" + StringProcessor.uniqueId() + "_" + timeStamp + ".zip");
-			uploadFile.save(tempFile);
-
-			try (InputStream input = new FileInputStream(tempFile)) {
-				// validate that the file is zip type
-				// .zip mime types  ---> application/zip, application/octet-stream, application/x-zip-compressed, multipart/x-zip
-				FileFormatCheck fileFormatCheck = new FileFormatCheck();
-				fileFormatCheck.setMimeType(uploadFile.getContentType());
-				fileFormatCheck.setFileFormat(fileFormat);
-				fileFormatCheck.setInput(input);
-
-				String formatError = service.getImportService().checkFormat(fileFormatCheck);
-				if (StringUtils.isNotBlank(formatError)) {
-					errors.put("uploadFile", formatError);
-				}
-			} catch (IOException ioe) {
-				throw ioe;
-			}
-		} catch (IOException ex) {
-			LOG.log(Level.FINE, "Unable to read file: " + uploadFile.getFileName(), ex);
-			errors.put("uploadFile", "Unable to read file: " + uploadFile.getFileName() + " Make sure the file in the proper format.");
-		} finally {
+			String filePath = FileSystemManager.getInstance().getDir(FileSystemManager.BULK_UPLOAD_DIR).toString() + "\\" + username + "\\" + timeStamp + "_" + StringProcessor.uniqueId() + ".zip";
 			try {
-				if (uploadFile != null) {
-					uploadFile.delete();
-				}
-			} catch (IOException ex) {
-				LOG.log(Level.WARNING, "Unable to remove temp upload file.", ex);
-			}
-		}
+				// save to file system under username
+				tempFile = new File(filePath);
+				uploadFile.save(tempFile);
 
-		// if successful send an email to SPOON support notifying of bulk upload and link to file path
+			} catch (IOException ex) {
+				LOG.log(Level.FINE, "Unable to read file: " + uploadFile.getFileName(), ex);
+				errors.put("uploadFile", "Unable to read file: " + uploadFile.getFileName() + " Make sure the file in the proper format.");
+			} finally {
+				try {
+					if (uploadFile != null) {
+						uploadFile.delete();
+					}
+				} catch (IOException ex) {
+					LOG.log(Level.WARNING, "Unable to remove temp upload file.", ex);
+				}
+			}
+
+			if(errors.isEmpty()){
+				Email email = MailManager.newEmail();
+				email.setSubject("SpoonSite bulk Upload");
+				email.setText("There is a new bulk upload to be reviewed at " + filePath);
+				email.addRecipient("", "support@spoonsite.com", Message.RecipientType.TO);
+
+				MailManager.send(email, true);
+			}
+		} else {
+			errors.put("uploadFile", "Uploaded file was not a zip file");
+		}
+		
 		return streamUploadResponse(errors);
 	}
